// src/app/lib/generated/models/RelationshipStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RelationshipStatus = {
    Girlfriend: 'girlfriend',
    Sexfriend: 'sexfriend',
    Stepmom: 'stepmom',
    Stepsister: 'stepsister',
    Colleague: 'colleague',
    SchoolMate: 'school_mate',
    Wife: 'wife',
    Mistress: 'mistress',
    Stranger: 'stranger',
    Friend: 'friend'
} as const;
export type RelationshipStatus = typeof RelationshipStatus[keyof typeof RelationshipStatus];


export function instanceOfRelationshipStatus(value: any): boolean {
    for (const key in RelationshipStatus) {
        if (Object.prototype.hasOwnProperty.call(RelationshipStatus, key)) {
            if (RelationshipStatus[key as keyof typeof RelationshipStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RelationshipStatusFromJSON(json: any): RelationshipStatus {
    return RelationshipStatusFromJSONTyped(json, false);
}

export function RelationshipStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): RelationshipStatus {
    return json as RelationshipStatus;
}

export function RelationshipStatusToJSON(value?: RelationshipStatus | null): any {
    return value as any;
}

export function RelationshipStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): RelationshipStatus {
    return value as RelationshipStatus;
}